[BE] [add] add interfaces + filter logic
