+++ backend/src/models/categoryInterfaces.ts
@@ -1,0 +1,17 @@
+import { IDishFE } from './dishInterfaces';
+export interface ICategoryFE {
+  foodGroup: string,
+  extraGroup: string
+}
+
+export interface ICategories {
+  name: string;
+  hitRate: number;
+  dishes: [IDishFE];
+}
+
+export interface ICategoryBE {
+  menuGroup: string,
+  foodGroup: string,
+  extraGroup: string
+}
