+++ backend/src/models/locationInterfaces.ts
@@ -1,0 +1,6 @@
+export interface ILocation {
+  streetName: string,
+  streetNumber: string,
+  postalCode: string,
+  country: string;
+}
