+++ backend/src/models/mealTypeInterfaces.ts
@@ -1,0 +1,5 @@
+export interface IMealType {
+  name: string;
+  id: number;
+  sortId: number;
+}
